Adjust format on step 1 for julios-salsa-copycat
Step 1 of the julios-salsa-copycat.html page was the only step with a
sentence structure. Removed periods from step 1.
There was also additional context to jalepeño seeds in regards to
spiciness of the salsa. This additional context was previously formatted
as a second sentence within the first step.
This commit removes the sentence structure and breaks the additional
note on jalepeño seeds as an unordered sub-list within the ordered list
of steps.