Adjust formatting on tortilla steps.
tortillas.html page had sentence structure where the other recipes did
not. This commit removes sentence structure.
Step 3 on tortillas was also phrased a bit strange, rearranged the
verbiage to flow better.
Added an unordered sub-list to step 8. In replacing the sentence
structure, step 8 no longer flowed as it had a second sentence for
additional context. This commit replaces the second sentence with an
unordered sub-list.