Adjust Formatting
Flexbox on index.html was not working to intention. This commit fixes
the causes of this.
Flexbox contents were supposed to be centered, but shifted to the right.
ul had 40px default padding which is now removed.
Flexbox contents were stacked on top of each other rather than side by
side. Flexbox was applied on the div .recipe-list. However, contents
were were in a ul within the div. a ul defaults to a block display.
This was cancelling out the flexbox. Moved display: flex property from .
recipe-list to the ul itself.
Lastly, I wanted the text for each recipe to appear layered over the
photos of the dishes. I applied a bottom margin of -44px to the photos
and added z-index values to the photos and the .caption class selector.
I also gave the .caption class selector a background to make the text
mor readable than it would have been over the photo alone.