Adjust formatting to prioritize flexboxes
Visually improved the recipe pages with flexboxes to display the recipe
photo side by side to the paragraph description when the screen is wide
enough to do so.
This commit also adds a flexbox to the recipe headers to account for an
even distribution of space between the header and new navigation links.
This commit also reorganizes the styles.css into 3 distinct sections
for general, index, and recipe pages. These css sections are notated
with comments. Where applicable, sections are ordered by page placement.