Make Index recipe photos react as a link
The Recipe photos on index.html were not acting as links to the recipes.
Only captions were acting as recipe links.
I found this to be caused by the a element being wrapped in the list
item element. I rearranged the list items to all be wrapped within
their respective a linking element.
This commit causes clicking the recipe photo to take you to the recipe.